feat(aws): add iam user allowed to assume agent
* add ability for iam users to assume the agent role

NOTE:

This should never be used in production environments. While this enables
an IAM user to be granted permission to assume the role, it does not
explicitly provide that grant to any given user. A policy must be
attached to a specific IAM user or IAM group to enable the role to be
assumed.